+++ serverless-test/src/other/m-import.ts
@@ -1,9 +1,13 @@
-import { response } from "@auto-docs/serverless";
-
 export const mImport = async () => {
   console.log("otherImport");
 
-  const myResponse = response(200, {});
+  const myResponse = {
+    statusCode: 200,
+    body: JSON.stringify({
+      message: "Hello from mImport!",
+      input: process.env,
+    }),
+  };
 
   return myResponse;
 };