+++ packages/serverless/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @auto-docs/serverless
 
+## 1.1.1
+
+### Patch Changes
+
+- add custom config
+
+## 1.1.0
+
+### Minor Changes
+
+- added build command
+
 ## 1.0.1
 
 ### Patch Changes